+++ examples/dev-server-ec2/main.tf
@@ -85,7 +85,7 @@ module "example_client_app" {
     mountPoints = []
     volumesFrom = []
   }]
-  retry_join = module.dev_consul_server.server_dns
+  retry_join = [module.dev_consul_server.server_dns]
 }
 
 # The server app is part of the service mesh. It's called
@@ -122,7 +122,7 @@ module "example_server_app" {
       }
     ]
   }]
-  retry_join = module.dev_consul_server.server_dns
+  retry_join = [module.dev_consul_server.server_dns]
 }
 
 resource "aws_lb" "example_client_app" {