+++ test/acceptance/tests/basic/terraform/basic-install/main.tf
@@ -201,7 +201,7 @@ EOT
       ]
     }
   ]
-  retry_join = module.consul_server.server_dns
+  retry_join = [module.consul_server.server_dns]
   upstreams = [
     {
       destination_name = "${var.server_service_name}_${var.suffix}"
@@ -245,7 +245,7 @@ module "test_server" {
     essential        = true
     logConfiguration = local.test_server_log_configuration
   }]
-  retry_join        = module.consul_server.server_dns
+  retry_join        = [module.consul_server.server_dns]
   log_configuration = local.test_server_log_configuration
   checks = [
     {