+++ test/acceptance/tests/basic/terraform/ca-cert-validate/main.tf
@@ -9,6 +9,6 @@ module "test_client" {
     name = "basic"
   }]
   outbound_only = true
-  retry_join    = "test"
+  retry_join    = ["test"]
   tls           = true
 }